refactor: Activity 파일들을 ui/activities로 이동
