+++ app/src/main/res/layout/activity_keyboard_setting.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.KeyboardSettingActivity">
+    tools:context=".ui.activities.KeyboardSettingActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
